Cherry picked all the ad filter related fixes from the master branch.
NOTE: that the adblock functionality still has known limitations in
kwebkitpart due to the fact that the source of a network resource request is
unknown. Currently there is no way to distinguish a request that originated as
a result of a <script> tag from one that came about as a result of a <style>
tag.

BUG:270194